+++ A1-2015/SystemA/FilterFramework.java
@@ -1,54 +1,18 @@
-/******************************************************************************************************************
-* File:FilterFramework.java
-* Course: 17655
-* Project: Assignment 1
-* Copyright: Copyright (c) 2003 Carnegie Mellon University
-* Versions:
-*	1.0 November 2008 - Initial rewrite of original assignment 1 (ajl).
-*
-* Description:
-*
-* This superclass defines a skeletal filter framework that defines a filter in terms of the input and output
-* ports. All filters must be defined in terms of this framework - that is, filters must extend this class
-* in order to be considered valid system filters. Filters as standalone threads until the inputport no longer
-* has any data - at which point the filter finishes up any work it has to do and then terminates.
-*
-* Parameters:
-*
-* InputReadPort:	This is the filter's input port. Essentially this port is connected to another filter's piped
-*					output steam. All filters connect to other filters by connecting their input ports to other
-*					filter's output ports. This is handled by the Connect() method.
-*
-* OutputWritePort:	This the filter's output port. Essentially the filter's job is to read data from the input port,
-*					perform some operation on the data, then write the transformed data on the output port.
-*
-* FilterFramework:  This is a reference to the filter that is connected to the instance filter's input port. This
-*					reference is to determine when the upstream filter has stopped sending data along the pipe.
-*
-* Internal Methods:
-*
-*	public void Connect( FilterFramework Filter )
-*	public byte ReadFilterInputPort()
-*	public void WriteFilterOutputPort(byte datum)
-*	public boolean EndOfInputStream()
-*
-******************************************************************************************************************/
-
 import java.io.*;
 
 public class FilterFramework extends Thread
 {
 	// Define filter input and output ports
 
-	private PipedInputStream InputReadPort = new PipedInputStream();
-	private PipedOutputStream OutputWritePort = new PipedOutputStream();
+	PipedInputStream InputReadPort = new PipedInputStream();
+	PipedOutputStream OutputWritePort = new PipedOutputStream();
 
 	// The following reference to a filter is used because java pipes are able to reliably
 	// detect broken pipes on the input port of the filter. This variable will point to
 	// the previous filter in the network and when it dies, we know that it has closed its
 	// output pipe and will send no more data.
 
-	private FilterFramework InputFilter;
+	public FilterFramework InputFilter;
 
 	/***************************************************************************
 	* InnerClass:: EndOfStreamExeception
@@ -65,7 +29,7 @@ public class FilterFramework extends Thread
 	****************************************************************************/
 
 	class EndOfStreamException extends Exception {
-		
+
 		static final long serialVersionUID = 0; // the version for serializing
 
 		EndOfStreamException () { super(); }
@@ -97,6 +61,7 @@ public class FilterFramework extends Thread
 			// Connect this filter's input to the upstream pipe's output stream
 
 			InputReadPort.connect( Filter.OutputWritePort );
+			System.out.println("Connected filter " + this.getName() + " - " + Filter.getName());
 			InputFilter = Filter;
 
 		} // try
@@ -108,6 +73,38 @@ public class FilterFramework extends Thread
 		} // catch
 
 	} // Connect
+
+	Measurement ReadMeasurement() throws EndOfStreamException {
+
+		try {
+			while (InputReadPort.available()==0 ) {
+				if (EndOfInputStream()) {
+					throw new EndOfStreamException("End of input stream reached");
+				}
+				sleep(250);
+			}
+		}
+		catch( EndOfStreamException Error ) {
+			throw Error;
+		}
+		catch( Exception Error ) {
+			System.out.println( "\n" + this.getName() + " Error in read port wait loop::" + Error );
+		}
+
+		/***********************************************************************
+		 * If at least one byte of data is available on the input
+		 * pipe we can read it. We read and write one byte to and from ports.
+		 ***********************************************************************/
+		try {
+			ObjectInputStream objectInputPort = new ObjectInputStream(InputReadPort);
+			Measurement datum = (Measurement)objectInputPort.readObject();
+			return datum;
+		}
+		catch( Exception Error ) {
+			System.out.println( "\n" + this.getName() + " Pipe read error::" + Error );
+			return null;
+		}
+	}
 
 	/***************************************************************************
 	* CONCRETE METHOD:: ReadFilterInputPort
@@ -240,19 +237,9 @@ public class FilterFramework extends Thread
 	*
 	****************************************************************************/
 
-	private boolean EndOfInputStream()
-	{
-		if (InputFilter.isAlive())
-		{
-			return false;
-
-		} else {
-
-			return true;
-
-		} // if
-
-	} // EndOfInputStream
+	boolean EndOfInputStream() {
+		return !InputFilter.isAlive();
+	}
 
 	/***************************************************************************
 	* CONCRETE METHOD:: ClosePorts
