+++ A1-2015/SystemA/SinkMeasurementMeasurementToStringCallback.java
