+++ A1-2015/SystemA/SinkMeasurementPrinter.java
